requirements.txt
Flask==2.2.2
numpy==1.23.5
opencv_python==4.6.0.66
Pillow==9.3.0
torch==1.13.0
